requirements.txt
langchain
langchain-google-genai
streamlit
sentence-transformers
chromadb